Starting to shape out a first approach
It's still unclear how far I'll be able to go with it,
though.

The basic idea is to collect all the nodes generated,
or referenced, when needed inside a List, and use
them for the operation.
This is still pretty crude and only tested on one binary
operator.

Also, the whole idea of functions with parameters
is kind of alien in Logix. The best way to emulate tihs
might be through Registers or Dynamic Variables.

Dynamic Variables sound cooler, but might be way
slower to use then simple registers.
When using simple registers, you can't really name
your inputs, though, and registers look in Logix is
wonky, at best (the actual storage node is stored
in a different slot, and have no Display, so Logix
arrows from elements writing into it look ilke they're
pointing to nothing...).