Now with simple function calls support
Still, user functions are not supported yet.
This will be the next step before trying to
tease about the project on Twitter.

I'm currently using registers for parameters,
but I still think I'll switch to dynamic
variables for parmeters, as it makes more sense
from a NeosVR point of view.

Though, using regiters for private functions
and dynamic variables for public functions might
also be a good idea.

Also, I'm really thinking about setting up
some kind of database (probably in JSON) to map
function calls to LogiX nodes.
The idea is that I'm still using a parser,
technically, which is not really a compiler.
So externals functions calls like "Color.ToHSV"
or fields access like "Time.current" are dealt
with the CSharpSyntaxWalker, and I have to
output something in correspondance.
Making a mapping table should make it easy to
configure the transpiler to different API, and
would also help if I do IL interpretation instead
of C# syntax interpretation.

Signed-off-by: Voyage <voyage@miouyouyou.fr>