Advancing with Impulse support...
At least, I was able to add support for default
outputs.
So the remaining parts is managing Run/Return,
and being able to actually use the returned value
correctly.

Then I still have to find a way to generate the
Dynamic variables. Probably through a COMPONENT
keyword.

Almost there for a simple demonstration, but
still a bit of work to do.

Signed-off-by: Voyage <voyage@miouyouyou.fr>