It works !
I am now able to parse a C# script, conver it
to a Logix Script similar (similar to what I
used for the Remote Logix Editor) and stream it
to NeosVR through WebSocket !

There's a still a LOT of things non implemented.
Actually, only what's required to convert a
very small script has been implemented.

However, user functions and returns are supported,
which was the biggest issue to me. Standard methods
just require some kind of translation database,
which will be the next goal here.
The idea is to have a database stating that, for
example :
Color.FromHSV(float,float,float) should be translated
as a "From HSV" node in Logix.
Of course, some functions will require a node
group to replicate such function.

Still, I'm happy to have finished this part, so
I'll tease about it on Twitter and NeosVR channels and
see the reactions.

Signed-off-by: Voyage <voyage@miouyouyou.fr>