Trying to add IF support
At first, I thought I'd able to one-shot this,
but this is more tricky than it looks.

Once an IF statement is encountered, it must
be executed if everything is fine. But whatever
the IF/ELSE result, the program continues.

If I only continue from the IF statement, the
IF statement becomes completely different.

If (a > 1.0f)
{
	b = 2;
}
DoThat();

is different from

If (a > 1.0f)
{
	b = 2;
	DoThat();
}

So, here, I need a Sequencer, to ensure that :
1. The IF statement is executed
2. The rest is executed normally

This... requires support for Outputs lists, to
connect the Impulse lines correctly, since "Sequence"
do not have simple named outputs.

I tried to avoid Outputs lists, but this is required here.
So next step :
- Impulse outputs lists...

Signed-off-by: Voyage <voyage@miouyouyou.fr>