Started to get support for basic branching
However, using the last affected node as the
current variable definition turned out to be
a bad idea, since this just doesn't work with
branching.

What happens is that when using this idea
with :

var a = 1;

if (check)
{
    a = 2;
}

Function(a);

When connecting 'a' to the function input,
first the literal '1' will be connected to
that input, then the literal '2' will be
also connected to the same input, erasing
the previous connection.

Which means that, in the end, only '2' will
be connected, making the branching completely
wrong.

So, next step :

* Variables as registers.
... Might go with that only when branching,
since this would imply a full write/read
sequence for each affectation of a local
variable which seems rather heavy...
...
Though optimizing too early might just break
for no reason...

Signed-off-by: Voyage <voyage@miouyouyou.fr>