Midway commit
Turns out that adding support for "local checkpoints"
require a LOT of modifications to make it run correctly.

And event with those, I'm starting to think that I
might need to merge RemoteLogix and SharpLogix somehow.
The reason being that I need an internal nodes
representation when trying, for example, to generate a
register to store the output of a generic node
(ReadDynamicVariable).

Unless I find a way to weasel out by abusing the auto
casting system within the Neos graphical side, but this
might require some very ugly Reflection use from the
plugin side...

So, yeah, in the end RemoteLogix HAS a representation
of the nodes. With this I can ping a node and get the
current outputs types, and all...
So combining both might not be a bad idea... However,
the issue is that I'm using CSharpSyntaxWalker here,
and I don't think there are similar C/C++ libraries
doing the same.
While Godot support C#, its support is less versatile
than GodotScript. For example, generating HTML5 apps
from Godot C# can quickly be a pain.

The only solution to use any language, would be to
create another SharpLogix clone that would focus on
IL parsing, instead of code parsing, which was still an
idea in the first place... It has a lot of upsides
(precompiled stuff, you're almost sure that the code is
valid unless the user REALLY want to mess with you, ...).

Still, now that I'm starting with C#, I'd still like
to provide basic support, and internal node representation
isn't the hardest thing to implement.
However, it's just that "I already done it", so I'd like
to avoid remaking it again.

Signed-off-by: Voyage <voyage@miouyouyou.fr>